Add UUID support
With this commit identifiying nodes by supplying a UUID is supported.
The UUID will be included in the merlin header, and the first packet
recieved from a node, will be used used to identify the socket
connection. UUID mode can be used on specific nodes only, and Merlin
operates as usual, when there are no nodes with UUIDs specified.

A new mon tool to generate a UUID is added.

Signed-off-by: Jacob Hansen <jhansen@op5.com>