Store cluster state in node struct
With this commit we store the cluster state in the merlin node struct.
We set the value when we find out the remote node has a wrong cluster
configuration. This allows us to send a CTRL_INACTIVE_CLUSTER instead of
CTRL_ACTIVE next a CTRL packet is sent. The remote node can then update
it's cluster configuration.

Signed-off-by: Jacob Hansen <jhansen@op5.com>